% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPercentMtRibo.R
\name{addPercentMtRibo}
\alias{addPercentMtRibo}
\title{Add percentage of mitochondrial and ribosomal transcripts.}
\usage{
addPercentMtRibo(object, organism, gene_nomenclature)
}
\arguments{
\item{object}{Seurat object.}

\item{organism}{Organism, can be either human ("hg") or mouse ("mm"). Genes need to annotated as gene symbol, e.g. MKI67 (human) / Mki67 (mouse).}

\item{gene_nomenclature}{Define if genes are saved by their name ('name') or ENSEMBL ID ('id').}
}
\description{
Get percentage of transcripts of gene list compared to all transcripts per cell.
}
\details{
Add percentage of mitochondrial and ribosomal transcripts.
}
\examples{
calculatePercentGenes(object = seurat, genes = gene_list)
}
\keyword{cerebro}
\keyword{seurat}
