% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateMarkerGenes.R
\name{annotateMarkerGenes}
\alias{annotateMarkerGenes}
\title{Annotate marker genes with EnrichR.}
\usage{
annotateMarkerGenes(object, column_sample = "sample",
  column_cluster = "cluster", adj_p_cutoff = 0.01, max_terms = 100)
}
\arguments{
\item{object}{Seurat object.}

\item{column_sample}{Column in object@meta.data that contains information about sample; defaults to "sample".}

\item{column_cluster}{Column in object@meta.data that contains information about cluster; defaults to "cluster".}

\item{adj_p_cutoff}{Cut-off for adjusted p-value of enriched pathways; defaults to 0.01,}

\item{max_terms}{Save only first n entries of each database; defaults to 100.}
}
\description{
This function uses the enrichR API to look for enriched pathways in marker gene sets of samples and clusters.
}
\examples{
annotateMarkerGenes(object = seurat)
}
\keyword{cerebro}
\keyword{seurat}
