% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performGeneSetEnrichmentAnalysis.R
\name{performGeneSetEnrichmentAnalysis}
\alias{performGeneSetEnrichmentAnalysis}
\title{Perform gene set enrichment analysis with GSVA.}
\usage{
performGeneSetEnrichmentAnalysis(object, column_sample = "sample",
  column_cluster = "cluster", thresh_p_val = 0.05,
  thresh_q_val = 0.1, gene_sets_to_test = "msigdb")
}
\arguments{
\item{object}{Seurat object. log-counts for analysis must be stored in
`object@data` (Seurat object older than v3) or `object@assays$RNA@data`
(Seurat object v3 or newer).}

\item{column_sample}{Column in object@meta.data that contains information
about sample; defaults to 'sample'.}

\item{column_cluster}{Column in object@meta.data that contains information
about cluster; defaults to 'cluster'.}

\item{thresh_p_val}{Threshold for p-value, defaults to 0.05.}

\item{thresh_q_val}{Threshold for q-value, defaults to 0.1.}

\item{gene_sets}{GMT file containing the gene sets to be tested. Defaults to
all gene sets stored in MSigDB [v7.0]:
http://software.broadinstitute.org/gsea/msigdb/index.jsp}
}
\description{
This function calculates enrichment scores, p- and q-value
statistics for provided gene sets for samples and clusters of cells in given
Seurat object using gene set variation analysis (GSVA).
}
\details{
Perform gene set enrichment analysis with GSVA.
}
\examples{
seurat <- performGeneSetEnrichmentAnalysis(
  object = seurat,
  column_sample = 'sample',
  column_cluster = 'cluster',
  thresh_p_val = 0.05,
  thresh_q_val = 0.1,
  gene_sets = 'msigdb'
)
}
\keyword{cerebro}
\keyword{seurat}
